Frontend: add performances

+++ TeacherTool_API/database/migrations/2018_10_18_141041_create_subjects_assessments_table.php
@@ -15,7 +15,7 @@ class CreateSubjectsAssessmentsTable extends Migration
     {
         Schema::create('subjects_assessments', function (Blueprint $table) {
             $table->increments('id');
-            $table->decimal('scale_factor');
+            $table->float('scale_factor');
             $table->char('type')->default('n');
             $table->integer('subject_id')->unsigned();
             $table->integer('assessment_id')->unsigned();
